pipelines/ic/function/requirements.txt
requests==2.24.0
google-cloud-storage~=1.44.0
tensorflow-cpu~=2.8.0
